+++ tests/e2e/urls.py
@@ -8,6 +8,7 @@ urlpatterns = [
     path("basic/", views.basic_view),
     path("elements/", views.elements_view),
     path("simple/", views.simple_view),
+    path("inline-errors/", views.inline_errors_view),
     path("builtin/", views.builtin_view),
     path("search-select/", views.search_select_view),
     path("multi-select/", views.multi_select_view),